Rename pour-via-escape fixtures tests and snapshots
