+++ fixtures/features/pour-via-escape/pour-via-escape03.fixture.tsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { AutoroutingPipelineDebugger } from "lib/testing/AutoroutingPipelineDebugger"
-import bugReportJson from "./bugreport53-1ce31b.json"
+import bugReportJson from "./pour-via-escape03.json"
 export default () => {
   return <AutoroutingPipelineDebugger srj={bugReportJson.simple_route_json} />
 }
