+++ fixtures/features/pour-via-escape/pour-via-escape04.fixture.tsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { AutoroutingPipelineDebugger } from "lib/testing/AutoroutingPipelineDebugger"
-import bugReportJson from "./bugreport50-da4a24.json"
+import bugReportJson from "./pour-via-escape04.json"
 export default () => {
   return <AutoroutingPipelineDebugger srj={bugReportJson.simple_route_json} />
 }
