+++ fixtures/features/pour-via-escape/pour-via-escape05.fixture.tsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { AutoroutingPipelineDebugger } from "lib/testing/AutoroutingPipelineDebugger"
-import bugReportJson from "./bugreport52-722ef1.json"
+import bugReportJson from "./pour-via-escape05.json"
 export default () => {
   return <AutoroutingPipelineDebugger srj={bugReportJson.simple_route_json} />
 }
