+++ tests/features/pour-via-escape/pour-via-escape05.test.ts
@@ -1,14 +1,14 @@
 import { expect, test } from "bun:test"
 import { AutoroutingPipelineSolver } from "lib"
-import bugReport from "../../fixtures/bug-reports/bugreport51-f1e2d3/bugreport51-f1e2d3.json" with {
+import bugReport from "../../../fixtures/features/pour-via-escape/pour-via-escape05.json" with {
   type: "json",
 }
 import type { SimpleRouteJson } from "lib/types"
-import { getLastStepSvg } from "../fixtures/getLastStepSvg"
+import { getLastStepSvg } from "../../fixtures/getLastStepSvg"
 
 const srj = bugReport.simple_route_json as SimpleRouteJson
 
-test("bugreport51-f1e2d3.json", () => {
+test.skip("pour-via-escape05.json", () => {
   expect(srj.layerCount).toBe(4)
   expect(
     srj.obstacles.some(
